requirements.txt
# WhispNote Requirements
# Core Streamlit app
streamlit>=1.28.0
streamlit-webrtc>=0.47.0

# AI/ML Models
openai-whisper>=20231117
torch>=1.13.0
transformers>=4.21.0
sentence-transformers>=2.2.0

# Keyword extraction
keybert>=0.7.0
scikit-learn>=1.1.0

# OCR
pytesseract>=0.3.10
easyocr>=1.7.0
Pillow>=9.0.0

# Document export
python-docx>=0.8.11
reportlab>=3.6.0

# Audio processing
librosa>=0.9.2
pydub>=0.25.1
soundfile>=0.10.3

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Utilities
pathlib2>=2.3.7
python-dateutil>=2.8.2

# Optional dependencies for enhanced functionality
# Uncomment if needed:
# tensorflow>=2.9.0  # For advanced ML models
# spacy>=3.4.0       # For better NLP
# fastapi>=0.95.0    # For API endpoints
# uvicorn>=0.18.0    # For API server

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0